Deleting temperature_test_freertos and putting its methods to temperature_test
