+++ test/src/CMakeLists.txt
@@ -1,7 +1,6 @@
 target_sources(
     ${PROJECT_NAME} PUBLIC
     temperature_test.cpp
-    temperature_test_freertos.cpp
     lora_test.cpp
 )
